Fixed Testing and Fixed NonLeaf Locate
Insert was not working because NonLeaf Locate was not working. Fixed it
so that if locate returns and error, then locateChildPtr returns the pid
of the last ptr in the file assuming that the key is greater than
everything in the node. Seems to work on test case of size 85.